requirements-dev.txt
-r requirements.txt
black
flake8
packaging
pre-commit
pytest
pytest-flask
reorder-python-imports
tox
